+++ src/median_log.h
@@ -3,11 +3,11 @@
 
 
 
-void MedianLog(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes, double *resultsSE);
-void MedianLog_noSE(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes);
+void MedianLog(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes, double *resultsSE);
+void MedianLog_noSE(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes);
 
-void medianlog(double *data, int rows, int cols, double *results, double *resultsSE);
-void medianlog_no_copy(double *data, int rows, int cols, double *results, double *resultsSE);
+void medianlog(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
+void medianlog_no_copy(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
